dayRate is now a double, allowing decimal time rate increases or reductions

+++ patches/tModLoader/Terraria/ModLoader/ModSystem.cs
@@ -220,10 +220,11 @@ namespace Terraria.ModLoader
 		public virtual void PostDrawTiles() { }
 
 		/// <summary>
-		/// Called after all other time calculations. Can be used to modify the speed at which time should progress per tick in seconds, along with the rate at which the world should update with it.
+		/// Called after all other time calculations.
+		/// Can be used to modify the speed at which time should progress per tick (measured in in-game minutes per real-life second), along with the rate at which the world should update with it (measured in ticks per frame).
 		/// You may want to consider Main.fastForwardTime and CreativePowerManager.Instance.GetPower<CreativePowers.FreezeTime>().Enabled here.
 		/// </summary>
-		public virtual void ModifyTimeRate(ref int timeRate, ref int tileUpdateRate) { }
+		public virtual void ModifyTimeRate(ref double timeRate, ref int tileUpdateRate) { }
 
 		/// <summary>
 		/// Allows you to save custom data for this system in the current world. Useful for things like saving world specific flags. For example, if your mod adds a boss and you want certain NPC to only spawn once it has been defeated, this is where you would store the information that that boss has been defeated in this world. Returns null by default.
